# frontend/openapi.json
{
	"components": {
		"schemas": {
...
			"CreatePostInputBody": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["https://example.com/schemas/CreatePostInputBody.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"content": { "description": "Post content", "type": "string" },
					"media_ids": {
						"description": "Media attachment IDs to include",
						"items": { "type": "string" },
						"type": ["array", "null"]
					},
					"random_delay_minutes": {
						"description": "Random delay in minutes (±N) to add for natural posting",
						"format": "int64",
						"type": "integer"
					},
					"scheduled_at": {
						"description": "Schedule time (ISO 8601). Omit for draft.",
						"format": "date-time",
						"type": "string"
					},
					"social_account_ids": {
						"description": "Social account IDs to publish to",
						"items": { "type": "string" },
						"type": ["array", "null"]
					},
					"workspace_id": { "description": "Target workspace ID", "type": "string" }
				},
				"required": ["workspace_id", "content", "social_account_ids"],
				"type": "object"
			},
...
			"DeletePostOutputBody": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["https://example.com/schemas/DeletePostOutputBody.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"message": { "description": "Success message", "type": "string" }
				},
				"required": ["message"],
				"type": "object"
			},
...
			"ErrorDetail": {
				"additionalProperties": false,
				"properties": {
					"location": {
						"description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
						"type": "string"
					},
					"message": { "description": "Error message text", "type": "string" },
					"value": { "description": "The value at the given location" }
				},
				"type": "object"
			},
			"ErrorModel": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["https://example.com/schemas/ErrorModel.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"detail": {
						"description": "A human-readable explanation specific to this occurrence of the problem.",
						"examples": ["Property foo is required but is missing."],
						"type": "string"
					},
					"errors": {
						"description": "Optional list of individual error details",
						"items": { "$ref": "#/components/schemas/ErrorDetail" },
						"type": ["array", "null"]
					},
					"instance": {
						"description": "A URI reference that identifies the specific occurrence of the problem.",
						"examples": ["https://example.com/error-log/abc123"],
						"format": "uri",
						"type": "string"
					},
					"status": {
						"description": "HTTP status code",
						"examples": [400],
						"format": "int64",
						"type": "integer"
					},
					"title": {
						"description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
						"examples": ["Bad Request"],
						"type": "string"
					},
					"type": {
						"default": "about:blank",
						"description": "A URI reference to human-readable documentation for the error.",
						"examples": ["https://example.com/errors/example"],
						"format": "uri",
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"PostDestinationResponse": {
				"additionalProperties": false,
				"properties": {
					"error_message": {
						"description": "Error message if publishing failed",
						"type": "string"
					},
					"platform": { "description": "Platform name", "type": "string" },
					"social_account_id": { "description": "Social account ID", "type": "string" },
					"status": { "description": "Destination status", "type": "string" }
				},
				"required": ["social_account_id", "platform", "status"],
				"type": "object"
			},
			"PostDetailResponse": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["https://example.com/schemas/PostDetailResponse.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"actual_run_at": {
						"description": "Actual run time after random delay (ISO 8601)",
						"type": "string"
					},
					"content": { "description": "Post content", "type": "string" },
					"created_at": { "description": "Creation time (ISO 8601)", "type": "string" },
					"created_by": { "description": "Creator user ID", "type": "string" },
					"destinations": {
						"description": "Post destinations",
						"items": { "$ref": "#/components/schemas/PostDestinationResponse" },
						"type": ["array", "null"]
					},
					"id": { "description": "Post ID", "type": "string" },
					"media": {
						"description": "Attached media",
						"items": { "$ref": "#/components/schemas/PostMediaResponse" },
						"type": ["array", "null"]
					},
					"random_delay_minutes": {
						"description": "Random delay in minutes (±N)",
						"format": "int64",
						"type": "integer"
					},
					"scheduled_at": { "description": "Scheduled time (ISO 8601)", "type": "string" },
					"status": { "description": "Post status", "type": "string" },
					"workspace_id": { "description": "Workspace ID", "type": "string" }
				},
				"required": [
					"id",
					"workspace_id",
					"created_by",
					"content",
					"status",
					"scheduled_at",
					"random_delay_minutes",
					"created_at"
				],
				"type": "object"
			},
			"PostMediaResponse": {
				"additionalProperties": false,
				"properties": {
					"alt_text": { "description": "Alt text for accessibility", "type": "string" },
					"display_order": { "description": "Display order", "format": "int64", "type": "integer" },
					"file_path": { "description": "File path for media URL", "type": "string" },
					"media_id": { "description": "Media ID", "type": "string" },
					"mime_type": { "description": "Media MIME type", "type": "string" }
				},
				"required": ["media_id", "display_order", "file_path", "mime_type", "alt_text"],
				"type": "object"
			},
			"PostResponse": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["https://example.com/schemas/PostResponse.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"actual_run_at": {
						"description": "Actual run time after random delay (ISO 8601)",
						"type": "string"
					},
					"content": { "description": "Post content", "type": "string" },
					"created_at": { "description": "Creation time (ISO 8601)", "type": "string" },
					"created_by": { "description": "Creator user ID", "type": "string" },
					"destinations": {
						"description": "Post destinations",
						"items": { "$ref": "#/components/schemas/PostDestinationResponse" },
						"type": ["array", "null"]
					},
					"id": { "description": "Post ID", "type": "string" },
					"media_ids": {
						"description": "Attached media IDs",
						"items": { "type": "string" },
						"type": ["array", "null"]
					},
					"random_delay_minutes": {
						"description": "Random delay in minutes (±N)",
						"format": "int64",
						"type": "integer"
					},
					"scheduled_at": { "description": "Scheduled time (ISO 8601)", "type": "string" },
					"status": {
						"description": "Post status (draft, scheduled, publishing, published, failed)",
						"type": "string"
					},
					"workspace_id": { "description": "Workspace ID", "type": "string" }
				},
				"required": [
					"id",
					"workspace_id",
					"created_by",
					"content",
					"status",
					"scheduled_at",
					"random_delay_minutes",
					"created_at"
				],
				"type": "object"
			},
...
			"UpdatePostInputBody": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["https://example.com/schemas/UpdatePostInputBody.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"content": { "description": "Post content", "type": "string" },
					"media_ids": {
						"description": "Media attachment IDs to include (replace all)",
						"items": { "type": "string" },
						"type": ["array", "null"]
					},
					"random_delay_minutes": {
						"description": "Random delay in minutes (±N) to add for natural posting",
						"format": "int64",
						"type": "integer"
					},
					"scheduled_at": {
						"description": "Schedule time (ISO 8601). Set to empty string to unschedule (make draft).",
						"type": "string"
					},
					"social_account_ids": {
						"description": "Social account IDs to publish to (replace all)",
						"items": { "type": "string" },
						"type": ["array", "null"]
					}
				},
				"type": "object"
			},
...
		}
	},
	"info": { "title": "OpenPost API", "version": "1.0.0" },
	"openapi": "3.1.0",
	"paths": {
...
		"/posts": {
			"get": {
				"operationId": "list-posts",
				"parameters": [
					{
						"description": "Filter by workspace ID",
						"explode": false,
						"in": "query",
						"name": "workspace_id",
						"schema": { "description": "Filter by workspace ID", "type": "string" }
					},
					{
						"description": "Filter by date (YYYY-MM-DD)",
						"explode": false,
						"in": "query",
						"name": "date",
						"schema": { "description": "Filter by date (YYYY-MM-DD)", "type": "string" }
					},
					{
						"description": "Filter by status (draft, scheduled, published, failed)",
						"explode": false,
						"in": "query",
						"name": "status",
						"schema": {
							"description": "Filter by status (draft, scheduled, published, failed)",
							"type": "string"
						}
					},
					{
						"description": "Limit number of results (default 50, max 200)",
						"explode": false,
						"in": "query",
						"name": "limit",
						"schema": {
							"description": "Limit number of results (default 50, max 200)",
							"format": "int64",
							"type": "integer"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"items": { "$ref": "#/components/schemas/PostResponse" },
									"type": ["array", "null"]
								}
							}
						},
						"description": "OK"
					},
					"default": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Error"
					}
				},
				"summary": "List posts for a workspace",
				"tags": ["Posts"]
			},
			"post": {
				"operationId": "create-post",
				"requestBody": {
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/CreatePostInputBody" } }
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/PostResponse" } }
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Bad Request"
					},
					"422": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Unprocessable Entity"
					},
					"500": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Internal Server Error"
					}
				},
				"summary": "Create a new post",
				"tags": ["Posts"]
			}
		},
...
		"/posts/{id}": {
			"delete": {
				"operationId": "delete-post",
				"parameters": [
					{
						"description": "Post ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "description": "Post ID", "type": "string" }
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/DeletePostOutputBody" }
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Bad Request"
					},
					"403": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Forbidden"
					},
					"404": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Not Found"
					},
					"422": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Unprocessable Entity"
					},
					"500": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Internal Server Error"
					}
				},
				"summary": "Delete a post",
				"tags": ["Posts"]
			},
			"get": {
				"operationId": "get-post",
				"parameters": [
					{
						"description": "Post ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "description": "Post ID", "type": "string" }
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/PostDetailResponse" }
							}
						},
						"description": "OK"
					},
					"404": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Not Found"
					},
					"422": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Unprocessable Entity"
					},
					"500": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Internal Server Error"
					}
				},
				"summary": "Get a single post",
				"tags": ["Posts"]
			},
			"patch": {
				"operationId": "update-post",
				"parameters": [
					{
						"description": "Post ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": { "description": "Post ID", "type": "string" }
					}
				],
				"requestBody": {
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/UpdatePostInputBody" } }
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/PostDetailResponse" }
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Bad Request"
					},
					"403": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Forbidden"
					},
					"404": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Not Found"
					},
					"422": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Unprocessable Entity"
					},
					"500": {
						"content": {
							"application/problem+json": {
								"schema": { "$ref": "#/components/schemas/ErrorModel" }
							}
						},
						"description": "Internal Server Error"
					}
				},
				"summary": "Update a post",
				"tags": ["Posts"]
			}
		},
...
	}
}